fixed bugs, running cascade well now for big file as well

+++ bloom-filter.c
@@ -115,8 +115,8 @@ void bloom_filter_free(BloomFilter *bloomfilter)
 
 void bloom_filter_insert(BloomFilter *bloomfilter, BloomFilterValue value)
 {
-	unsigned int hash;
-	unsigned int subhash;
+	long long hash;
+	long long subhash;
 	long long index; //yeinhorn
 	unsigned int i;
 	unsigned char b;
@@ -149,8 +149,8 @@ void bloom_filter_insert(BloomFilter *bloomfilter, BloomFilterValue value)
 
 int bloom_filter_query(BloomFilter *bloomfilter, BloomFilterValue value)
 {
-	unsigned int hash;
-	unsigned int subhash;
+	long long hash;
+	long long subhash;
 	long long index;
 	unsigned int i;
 	unsigned char b;
